version 0.940
Bug fix: Was getting logcat message saying that thread had not called Looper.prepare (when refresh (vitals) was pressed).  So added Looper.prepare() on the thread that was created - this fixed the message - Also added Looper.loop() as this seems to be a requirement from what I've read - not sure if that is actually the case for what I'm doing.

  Looper.prepare();
                Server.updateServerConnectionStatus(getApplicationContext());
                Message message = handler.obtainMessage();
                message.what = RESUME_AND_DISPLAY_REFRESH_MESSAGE;
                message.sendToTarget();
                Looper.loop();