versionName "0.948"
bug fix 1:
Android API 15 gave error when no memory sim.  Fixed by using  appDataFolder = context.getFilesDir(); if  Environment.getExternalStorageState(); doesn't work
This will mean that it will be harder for another app to get the files off the phone as they will be in the private area

bug fix 2:
Android API 15 gave error when moving from main screen to setup screen.  Error was complaining that the RelativeLayout did not have an id so code could not return to the view.  Gave it an id in the activity_main.xml file