versionName "0.959"
Logger gave error on one phone - so reverted logging to Android Log class - but it turns out the phone had a fault with internal storage!!
But will leave code using Log
Added code to look for micro sd memory, and it will use that if it is present - this allows the broken phone to work, but also now means that sd memory can be used when no data connection is available - user does not have to do any configuration.  BUT, this needs to be tested on a range of OS versons.