versionName "0.960"
Added checkbox (No network connection) to indicate that app will be used offline.  This boolean will be used to allow test recordings even when no network connection.
Also, can use this value to not bother turning aerials on if we know there isn't going to be a connection

Next commit I'll change variable /checkbox name to be 'offline mode'