versionName "0.965"
Have been working on issue of GPS update causing network / app to crash.  Might be due to GPSLocationListener calling resume on SetupActivity.  I've stopped it calling resume which may have fixed the issue, but now the GPS coordinates on the SetUp page are not updated when the GPS location is found.  So, next I'll look at using a localbroadcast receiver to allow communiction between the GPSLocationListener and the SetupActivity - 'just' need to copy the same pattern as I use in MainActivity.