Deleted Old Tests.
Removed signingConfig from buildl.gradle (will just rely on Android Studio's popup dialog asking for location of keystore and password).  Note if choose for Android Studio to 'remember'   according to  (https://stackoverflow.com/questions/18701932/how-can-i-retrieve-a-saved-keystore-password-from-android-studio) "the password is saved in Android Studio 3.1 seems to use the OS keychain in Ubuntu 18.04," - not in a file in the project.