#31 Fix email address already being used error unknown.
Did a lot of refactoring, but also fixed the bug. Server returned valid data but we were checking strings exactly and it looks like the format changed.
Modified the error text to use the message directly so it will still work even in new API versions too.

Simplified a lot of logic and moved most of the POST calls to OkHttp3 instead of raw calls as they're much easier to work with.
This allowed a lot of logic to be refactored out/simplified too.