After phone connects to the server and before first upload starts, now check to see if user has pressed cancel upload button.
Cancel button no longer asks user if they are sure - as this was just wasting time if the user really really want to stop the uploading.

Pressing the cancel button immediately disables itself and sets message to inform user that uploading is being stopped.

Bug fix: If uploading is already occurring then code returns and an incorrect message about recording in progress was broadcast.  I was going to change the message to say 'Already Uploading', but I don't see a need for any message so removed it, and left a log statement.