Large refactor to improve the message broadcasting system.
There are two main things targeted here, reducing the code duplication in sending messages, and adding Enums to reduce copy paste errors

Firstly, I've refactored a lot of the broadcastMessage sending code to use more concise methods, reducing a lot of code duplication by creating the MessageHelper class and its relevant broadcastMessage methods.

Second, reworked the way messages are handled in general to make bugs/errors harder to happen.
This is done by enforcing the Action messages to use an Action class, and their possible messages to be defined in an Enum.
Because of this, classes must use the provided action/message types and you can't just typo something and miss it.
This also allows switch statements to be used which is easier to follow than the if statements of before.

Most of the smart stuff here is in MessageHelper, which provides ways to create/register/unregister the message handlers, as well as the functionality to broadcast. This class basically adds a wrapper to enforce the Action class/Enums to be used.