Updated tests to inherit from a base test class that manages all the standard setup/teardown to reduce code duplication. Made the tests fully restart/reset on each test case which makes them fully independent, they shouldn't rely on any initial state anymore.
Updated idlingResources to be incremented/decremented in the Server class mostly, to prevent issues with message handling. This has the downside that there might be slight gaps between idlingResources (e.g. sign in -> get groups). Doesn't seem to make the tests fail, but may be an issue if we want to remove sleeps.

Renamed a couple of things to be more clear/specific (disable -> disableAutomaticRecording; settingsForTestServerEnabled -> veryAdvancedSettingsEnabled)

Added extra library to allow "scroll" instead of "swipe" which is far more consistent for navigation.

Made preferences fully save/restore between each test so running tests on your real device leaves its settings consistent with how it was before.